% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_get_mapping_vector.R
\name{get_mapping_vector}
\alias{get_mapping_vector}
\title{Get a Named Vector from a Two-Column TSV}
\usage{
get_mapping_vector(tsv_file, has_header = TRUE)
}
\arguments{
\item{tsv_file}{Character. Path to the TSV file.}

\item{has_header}{Boolean. TRUE if the file has a header. Default is TRUE.}
}
\value{
A named vector.
}
\description{
Reads a two-column TSV file (e.g., mapping barcodes to sample names or
cluster IDs) and converts it into a named vector. The first column is
used as names, and the second column as values.
}
